base64

+++ ver1/panel/views/institucion/index.php
@@ -16,12 +16,22 @@
     <?php foreach ($data as $institucion): ?>
     <tr>
       <th scope="row"><?php echo $institucion['id_institucion']; ?></th>
-      <td><img src="../img/institucion/<?php echo $institucion['logotipo']; ?>" width="75" height="75" class="rounded-circle" alt="logo"></td>
-      <td><?php echo $institucion['institucion']; ?></td>
+      <td>
+        <?php 
+          $logotipo = $institucion['logotipo'];
+          if ($logotipo && strpos($logotipo, 'data:image/') === 0) {
+              $srcImagen = $logotipo;
+          } else if ($logotipo) {
+              $srcImagen = "../img/institucion/" . $logotipo;
+          }
+        ?>
+        <img src="<?php echo $srcImagen; ?>" width="75" height="75" class="rounded-circle" alt="logo" style="object-fit: cover;">
+      </td>
+      <td><?php echo htmlspecialchars($institucion['institucion']); ?></td>
       <td>
         <div class="btn-group" role="group" aria-label="Basic mixed styles example">
             <a href="institucion.php?action=update&id=<?php echo $institucion['id_institucion']; ?>" class="btn btn-warning">Editar</a>
-            <a href="institucion.php?action=delete&id=<?php echo $institucion['id_institucion']; ?>" class="btn btn-danger">Eliminar</a>
+            <a href="institucion.php?action=delete&id=<?php echo $institucion['id_institucion']; ?>" class="btn btn-danger" onclick="return confirm('¿Está seguro de eliminar esta institución?')">Eliminar</a>
         </div>
       </td>
     </tr>
